+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteDefinitions.cs
@@ -227,8 +227,8 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="style">Border style.</param>
         /// <param name="state">Palette value should be applicable to this state.</param>
-        /// <returns>Integer rounding.</returns>
-        int GetBorderRounding(PaletteBorderStyle style, PaletteState state);
+        /// <returns>Float rounding.</returns>
+        float GetBorderRounding(PaletteBorderStyle style, PaletteState state);
 
         /// <summary>
         /// Gets a border image.
@@ -1222,7 +1222,7 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="state">Palette value should be applicable to this state.</param>
         /// <returns>Integer rounding.</returns>
-        int GetBorderRounding(PaletteState state);
+        float GetBorderRounding(PaletteState state);
 
         /// <summary>
         /// Gets a border image.
