+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteRedirect/PaletteRedirectBorderEdge.cs
@@ -198,8 +198,8 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="style">Border style.</param>
         /// <param name="state">Palette value should be applicable to this state.</param>
-        /// <returns>Integer rounding.</returns>
-        public override int GetBorderRounding(PaletteBorderStyle style, PaletteState state)
+        /// <returns>Float rounding.</returns>
+        public override float GetBorderRounding(PaletteBorderStyle style, PaletteState state)
         {
             return Target.GetBorderRounding(style, state);
         }
@@ -241,7 +241,7 @@ namespace Krypton.Toolkit
             PaletteBorderEdge inherit = GetInherit(state);
 
             return inherit?.GetBackImageAlign(state) ?? Target.GetBorderImageAlign(style, state);
-        }    
+        }
         #endregion
 
         #region Implementation
